requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
PyPDF2
sentence_transformers
numpy
FAISS-cpu
python-multipart
groq
streamlit